RDKBWIFI-408: SteerWiFiBackhaul (parse only), ChannelScanRequest and Disassociate methods for TR-181 added
Reason for change: Getters, setters and methods of WFA datamodel parameters are needed for TR-181.
Test Procedure: (RDKB) rbuscli may be used to get parameters. Some examples:
  rbuscli method_values "Device.WiFi.DataElements.Network.Device.1.MultiAPDevice.Backhaul.SteerWiFiBackhaul()" TargetBSS string "81:12:13:14:15:16" TimeOut uint32 1000
  rbuscli method_values "Device.WiFi.DataElements.Network.Device.1.Radio.1.ChannelScanRequest()" OpClass uint32 81 ChannelList string "6,9"
  rbuscli method_values "Device.WiFi.DataElements.Network.Device.1.Radio.1.BSS.1.STA.1.MultiAPSTA.Disassociate()" DisassociationTimer uint32 3 ReasonCode uint32 5
Risks: Low

+++ src/ctrl/tr_181/wfa_data_model/tr_181_helper.cpp
@@ -225,17 +225,17 @@ bool tr_181_t::tr181_get_prop_int(const bus_data_prop_t *prop, int *value)
     }
 
     if (prop->value.data_type == bus_data_type_int8) {
-        *value = prop->value.raw_data.i8;
+        *value = static_cast<int>(prop->value.raw_data.i8);
     } else if (prop->value.data_type == bus_data_type_uint8) {
-        *value = prop->value.raw_data.u8;
+        *value = static_cast<int>(prop->value.raw_data.u8);
     } else if (prop->value.data_type == bus_data_type_int16) {
-        *value = prop->value.raw_data.i16;
+        *value = static_cast<int>(prop->value.raw_data.i16);
     } else if (prop->value.data_type == bus_data_type_uint16) {
-        *value = prop->value.raw_data.u16;
+        *value = static_cast<int>(prop->value.raw_data.u16);
     } else if (prop->value.data_type == bus_data_type_int32) {
         *value = prop->value.raw_data.i32;
     } else if (prop->value.data_type == bus_data_type_uint32) {
-        *value = prop->value.raw_data.u32;
+        *value = static_cast<int>(prop->value.raw_data.u32);
     } else {
         return false;
     }
